# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hiringo API Service.",
        "title": "Hiringo API",
        "contact": {
            "name": "Emin Muhammadi",
            "url": "https://github.com/softech-craftsman-webapp/hiringo",
            "email": "admin@hiringo.tech"
        },
        "version": "1.0"
    },
    "host": "main-api.hiringo.tech",
    "basePath": "/",
    "paths": {
...
        "/ratings/my": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get user ratings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rating"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/view.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/view.RatingView"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    }
                }
            }
        },
...
        "/ratings/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get rating Details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rating"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rating id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/view.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/view.RatingView"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    }
                }
            }
        },
        "/statistics/category": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Category statistics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/view.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/view.CategoryStatistics"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    }
                }
            }
        },
        "/statistics/job": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Job statistics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/view.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/view.JobStatistics"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/view.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "view.CategoryStatistics": {
            "type": "object",
            "properties": {
                "popular_item": {
                    "$ref": "#/definitions/view.CategoryView"
                },
                "time": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "view.CategoryView": {
            "type": "object",
            "properties": {
                "created_by_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "view.JobStatistics": {
            "type": "object",
            "properties": {
                "latest_job": {
                    "$ref": "#/definitions/view.JobView"
                },
                "time": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "user_job_count": {
                    "type": "integer"
                }
            }
        },
        "view.JobView": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "is_contract_signed": {
                    "type": "boolean"
                },
                "is_equipment_required": {
                    "type": "boolean"
                },
                "is_premium": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
...
        "view.RatingView": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "contract_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "submitted_by_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "view.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "payload": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}